Ensure extension aliases are used for nova based services
This commit fixes the extension checker in verify-tempest-config to
ensure that nova and nova based services are using the extension
aliases instead of the extension names. The extension names are meant
to be human readable and are not the best way to identify it. This is
what the tempest code is also expecting, however in some cases the
name was still being used by the verify tool, which was incorrect. As
part of this it was necessary to fix several other places where name
was being used instead of alias, or other extension list formatting
response checking was incorrect.

Change-Id: I35f5db655f065f29760515a26db1fb508d229393