Remove network debug
This commit removes all of the network debugging gorp usage from
tempest. These were added as a crutch over a year ago to enable
real neutron testing in the gate, which was impossible to debug
at the time. This debug info is an inherent layer violation in tempest
which is only supposed to be black-box and not care about any of the
underlying internals. It also assumes a single host environment and
that tempest is running on that machine, which is mostly specific to
the gate and devtest setups. Additionally the amount of output on a
failure is staggering to the point where keeping this around is
counter-productive. If we are unable to debug issues at this point
neutron needs to improve their logging so people can figure out what
failed.

Change-Id: Ia1ed4ab3cd2e4941fb072b3390f4859420989e74