Actually attach a volume to an instance before taking snapshot
In the test test_snapshot_create_with_volume_in_use, the test calls Cinder
"os-attach" for attaching a volume. The "os-attach" to tell Cinder the
volume is attached, but the API doesn't actually attach the volume to an
instance.(Only update volume status in DB)

This is not right test case for taking a snapshot with in-use volume.
In this test, Nova "os-volume_attachment" should be called for volume
attachment.

Also, some Cinder drivers fails assisted snapshot due to this problem.
In order to perform the snapshot properly, this fix is needed.

Closes-Bug #1401110
Change-Id: Ib31e351fe7c3d27824241cf142c213eae287483f