Remove Nova v3 API clients
Nova v3 API has disappeared, and Tempest isn't testing the API on the
gate now. In addition, Nova team is creating a new REST API "Nova v2.1
API + microversions" and the interfaces are different from v3 API.
So it is not necessary to keep Nova v3 API tests in Tempest.
This patch removes the clients.

Change-Id: I90f48b3adcd94435320de965a3db21eefaa97cda