Remove Nova v3 API schemas
Nova v3 API has disappeared, and Tempest isn't testing the API on the
gate now. In addition, Nova team is creating a new REST API "Nova v2.1
API + microversions" and the interfaces are different from v3 API.
So it is not necessary to keep Nova v3 API tests in Tempest.
This patch removes the API schemas.

Change-Id: Icf56194e855452e580e6a126f8477ebbd8d6952d