Use isolated creds for dashboard scenario
This commit modifies the dashboard basic ops scenario to use isolated
creds allocated during the setupClass. Previously this test used hard
coded values from the config file. But, moving forward those options
will be deprecated in favor of the test-accounts file. In preparation
for that shift this test needs to be modified to use the credentials
provided by either tenant isolation or test accounts.

Change-Id: I92d5809ef8c46350684ebbed7269cd6df04f39da