Remove Nova v3 API config
Nova v3 API has disappeared, and Tempest isn't testing the API on the
gate now. In addition, Nova team is creating a new REST API "Nova v2.1
API + microversions" and the interfaces are different from v3 API.
So it is not necessary to keep Nova v3 API tests in Tempest.
This patch removes the API config.

Change-Id: I6c0dc4ff1b05dcf29aae2937104dacf55384d12f