Change v3 identity client methods to return one value
Tests were updated. Although DataGenerator had already been modified to
treat v2/v3 differently there were a few methods in BaseIdentityAdminTest
that were not. The required methods are now overridden in the V3 subclass.
The V2/V3 separation here could be cleaned up a bit more but not in this
patch which is already large.

Partially implements: blueprint clients-return-one-value

Change-Id: Ie925b17fe352a7f942d413c32f156137e0cba7e4