Handle HTTP 415 in rest_client
If HTTP 415 ("Unsupported Media Type") is returned then,
rest_client responds with an "UnexpectedResponseCode" exception.

Tempest always send request with supported media type and that's
why 415 did not happen.
But as rest_client is on the way to be part of tempest-lib,
its better to handle 415 and raise appropriate error.

This patch handle HTTP 415 error and raise 'InvalidContentType'.

Change-Id: I72690aab7afd8e68f054f31315a8b1576a7ec9c4
Closes-Bug: #1407140