Add image client build timeout config option
This patch Ibc4307b406b71db258975da90a39c1f51ad27e12 changed
the RestClient so that it does not use the
CONF.compute.build_timeout variable anymore.

For this reason, there is no way to specify a timeout for the image
client, which inherits the hardcoded 60 seconds timeout.

This patch introduces a config option for this. For consistency,
a config option was added for build interval as well.

Change-Id: I0ac5ead60c4dc5a0c04feee513ae299ca14bd38e
Partial-bug: #1394519
Closes-bug: #1407943