Unskip and Add timeout method to dns check scenario
In test_subnet_details, Tempest checks the effect of alternative dns
server which is updated just before the check.
However, sometimes the check failed because the update could not make
it before the check. This reason is not concrete yet, but this problem
happens three times in 60 minutes now.
So we need to solve this problem soon for smooth development, and this
patch adds timeout method for alternative dns part.

Change-Id: I0cb98ff401b526c24388ea440e5629a8ebc925eb
Closes-Bug: #1412325