Remove inheritance from baremetal client
baremetal client has complicated inheritance because Tempest tested
XML tests and the base part of the client was necessary for both XML
and JSON. However, XML tests have been already removed from Tempest
and the inheritance is unnecessary and the code is complicated.
The this patch removes the inheritance and renames the file name of
the client for fittig to the other clients' names.

Change-Id: I0a798fb146edd53b7a61bf30689b93cbd90ddac0