Add sec-grp-rule for ipv6-icmp to enable ping6 between instances
The tempest testcase test_network_v6 checks if an ipv6 ping
between instance is possible. But therefore no security-group-
rule is being set up, to allow such a ping. The test fails.

This issue does not apply to a test that runs with the cirros
image, as cirros does not have ipv6 utils installed. There is
a if clause in the testcase, that skips this ping if the util
ping6 is not available.

Change-Id: Ia3fbe771cde8893e9eb3fc43b187435bd0dd1dbf
Closes-Bug: #1417584