Fix compute baremetal_nodes API schema
The Nova os-baremetal-nodes endpoint is effectively a proxy for the Ironic
API.  The API schema on the tempest side is too strict, does not reflect
the documented Nova responses and does not reflect the documented Ironic
responses.  Specifically, the node resource properties may be strings or
integers and Nova documents them currently as strings.  This updates the
Tempest schema to accept both.

Closes-bug: #1422832

Change-Id: I57e385014799aa6333a74a30696dd2846efdbc68