Omitting microseconds when creating a trust
In some cases the expiry time may be rounded up because of microseconds.
For example, we have the following expiry time for a trust:

    2015-02-17T17:34:01.907051Z.

However, if we make a GET request on the trust, the response may contain
the time rounded up to

    2015-02-17T17:34:02.000000Z.

That is why we should omit microseconds when creating a trust.

Closes-Bug: #1422869

Change-Id: I6664c4159e682a5388717297c8d49eddebad3f04