Use ssl conf settings for raw verify_tempest_config requests
This commit enables the use of the ssl settings from the tempest
config file for making raw requests from verify_tempest_config.
Previously, there were a set of calls to get the available
api_versions from an endpoint, which is just done by making a raw get
on an unversioned endpoint, however these neglected to respect the
ssl opts from the config file.

Change-Id: I397e75311578cef8ff0710c7040984ea8adb2239
Closes-Bug: #1420280