Fix baremetal node property keys
The baremetal_client is creating test nodes with property keys that
don't match what the nova os-baremetal-nodes extension is pulling out of
the node properties from the ironic client. This results in KeyErrors
when this test runs in parallel to other baremetal tests that were added
to Tempest on 2/16 with commit ac0879accd2a451c54ed935297a6f8e0a8aaf25a.

Closes-Bug: #1423427

Co-authored-by: Adam Gandelman <adamg@ubuntu.com>

Change-Id: I3593f58de5088d2442fddd52c6a11d055bb4f6a0