Add tempest config guide doc
This commit adds a new doc page to the tempest documentation to
provide a guide to configuring tempest. It currently contains one
section to elaborate on the credentials and auth options in the config
file. This doc will be expanded in the future to add details around
other pieces of tempest configuration.

Change-Id: Iaf7cefc6a5e508437e4b4620ba5bfaa221f3611b