Replaces hardcoded volume size by configured value
The volume_create methods used a hardcoded volume_size of 1GB. The
configured volume_size wasn't used on each place. If one used a 
configured image_ref with a volume_size > 1GB this would result in an
error.
This patch replaces all occurrences of the hardcoded volume_size by
the configured value.

Closes-Bug: 1289118
Change-Id: I99cf80e4078eec6a5021793b664b6cfcf02d2988