Fix check for FloatingIP status before check connection
This patch changes the order in which the floatingIP
status is being checked in all the network_basic and
network_advanced tests.

The check for floatingip status should be done before
your check for public network connectivity to the VM.

This would give the real details for the test case if
it was related to floatingip setup or something in the
test environment.

Change-Id: Ie606ba5194fe87047d553056ba0a82bcf8830ac0
Partial-Bug: #1426764