Fix isolated cred cleanup for cli tests
This commit renames the isolated_creds object initialized in the
setupclass for the cli tests. Previously it was using the var cred_prov
which prevented the clear_isolated_creds classmethod from detecting
isolated credentials had been provisioned. This meant that cleanup
was never run which would mean any credentials created or reserved
wouldn't be deleted or unreserved.

Change-Id: I56c270dd719338527d81da9cbd9857f54b76f51a