Make scenario tests requiring admin skip if no admin creds
Some scenario tests will fail if there are no admin creds. This commit removes
the creation of an admin client from the setup of the ScenarioTest class
and puts that code in the subclasses that need it, along with a check to skip
if no admin creds are available.

One such class was the unused OrchestrationScenarioTest which was removed.
Future heat scenario tests will go in the heat project.

Change-Id: Ib0b82510200e927a391576686e0a78c4c4b6cd34