Switch run_tempest.sh to use subunit-trace
This commit switches the run_tempest.sh script to use subunit-trace
instead of the colorizer output filter. Since the colorizer isn't used
anymore it is removed.

The colorizer output has several issues. Firstly the color is actually
meaningless. It's based on arbitrary run times for the tests but don't
actually indicate anything useful. Additionally, it lacks a lot of the
debugging information that subunit-trace provides. Things like skip
messages, the worker breakdown, etc. are missing with the colorizor
output. The last aspect is that switching to use subunit-trace will
provide a consistent output format for running tempest whether you
use tox or run_tempest.sh.

Change-Id: I26e9c45f8e674bdb889b6bc65ad30c4abf6e7655