Add test for the various config cases of is_admin_available()
This function is used to determine if admin credentials are available but
there are many different ways to configure admin in tempest. The admin
must be available if tenant isolation is enabled, otherwise can be specified
with either the admin values in the identity section of tempest.conf or by
adding an admin role or type to a user in the test accounts file.
Change-Id: Ibb82d1008ebace7f86a857761673fe9ab496a7ec