Distinguish between luks and cryptsetup volume types
TestEncryptedCinderVolumes was always creating the volume type in cinder
with a random name but 'luks' was always in the name since it's
hard-coded, which makes debugging failures in the cryptsetup test
difficult when you're looking at nova-compute logs.

Change the helper method to take a parameter for the volume type name
so we can distinguish between luks and cryptsetup.

Related-Bug: #1327218

Change-Id: If607ba6d389e6a482cfb1eb3af4fe895ed00f8db