javelin: don't destroy floating IP resource
The code to destroy floating IP resources is broken today and needs more
work.
To bring back the CI again, let's drop this code and add a TODO so we
can work on it later once CI is back and we also have a new job testing
Grenade on Neutron.

Change-Id: I8cf7fb3aa7e361d603716afbd3cf98827151f094