Make L3AgentSchedulerTestJSON DVR-aware
test_add_list_remove_router_on_l3_agent fails when testing with DVR
routers as distributed routers follow a different agent binding logic.

Due to this failure, performance improvements like e99f6e00 were reverted.
This called for a revision of the test case. This change ensures that,
when needed, the test provisions extra resources so that the binding logic
can be successful.

Closes-Bug: #1433349

Co-Authored-By: Armando Migliaccio <armamig@gmail.com>

Change-Id: I15bbf16fd4378c6431e9da8942d0968e7a012a91