Prevent failures when running as non-admin
In some cases a skip method was needed to check for non-admin. In other cases
there was already a try/except to handle on-admin but it needed to be changed
to use is_admin_available. NegativeAutoTest was changed to push creation of
an admin client to the point where the need has been established.

Also fixed some skip_checks methods to call super first.

Change-Id: Ie58135285a4c95c2eed9c0462c4e76af3ad5d97e