Fix external connectivity check for dualstack
This patch modifies the _check_network_external_connectivity()
method in test_network_basic_ops such that it asserts there exists
one IPv4 subnet (instead of one subnet of any address family) and such
that it tests external connectivity to only the IPv4 subnet (in the
case of a dualstack external network).

The external connectivity test currently relies on SSH'ing to an
instance through its floating IP and pinging the external gateway IP.
To test IPv6 external connectivity, this test would have to be
modified to SSH to an instance's IPv6 GUA address and use the ping6
utility to ping the external IPv6 gateway address. That enhancement is
outside the scope of this patch and can be implemented in the future.

Change-Id: I0f9ea98cb84aa72cb1505fb9ff8ac61561cc1376
Closes-Bug: #1440226
Partially-implements: blueprint multiple-ipv6-prefixes