Handle fixed_network edge cases gracefully
When we added support for using the fixed_network_name config option to
specify the network to boot with in a multi-network env a couple of
configuration edge cases were not taken into account. First the case
of misconfiguration was not handled at all this would cause an ugly
IndexError exception to be raised because no matches were found for
the name specified in config. The other was since the default config
option was set to 'private' and fixed network name is always used when
configured the default for single network environments broke if the
single network. This commit addresses these by removing the default
value for fixed_network_name (and making the help more clear) and
having fixed_network_name handle the misconfiguration case more
clearly by raising an InvalidConfiguration exception.

Change-Id: I06ac0605a1a7e35d1af9a93a3bfc387a78f8be1c
Closes-Bug: #1437328