Limit tempest.config's use of cfg.CONF
This change supports the resuse of tempest's config machinery by other
projects such as Neutron.  Use of cfg.CONF can be replaced with a
local ConfigOpts instance by changing a single line (preventing
conflict with the consuming project's config), which can be easily
maintained by a copy script.  This ensures that a project that uses
the cfg.CONF global won't run into conflicts with tempest's
configuration.

Change-Id: Ia2cdd54af8c06d3a38a62183c1f03cd317c3972d