Wait for complete deletion of volume snapshot before deleting volume
In file tempest/scenario/test_volume_boot_pattern.py there 
is a method for creating snapshots from volumes. The method 
is _create_snapshot_from_volume(...). The thing is that method
_create_snapshot_from_volume(...) uses method 
self.addCleanup_with_wait(...) (that does not wait for complete 
resource deletion right after sending the request to delete the 
resource) in order to delete created volume snapshot after the test
finished. But volumes cannot be deleted when there are outstanding 
snapshots. Therefore we should wait for complete deletion of 
the volume snapshot before we start to delete the volume.

Closes-Bug: #1441563

Change-Id: I11207c3ac9296389c5e195b2296fc47e65a7486c