Add router to resources section of yaml output
network and router are both top-level network
resources and router should be represented in
the yaml file both for clarity and also having
router name exposed in yaml file is useful for
building white list for tempest cleanup.

In addition horizon also lists networks and routers
as top-level network objects. Adding router to the
test account output aligns with this existing pattern.

Change-Id: Iddebba793a978281f698ea10c167dc0a8765d785