Set default value for tempest log file
This commit changes the default value for tempest's logging output.
Since tempest is intended to be run solely through a test runner
logging to stdout can get complicated since tempest doesn't really
own stdout. This patch changes the default to log to tempest.log in
the cwd where tempest is run which should be a better behavior for
running tempest.

Change-Id: I1017a1714181a78b5da9c47e811db63b4edce14a