Cleanup fixed_name logic in test_list_server_filters
This commit cleans up the logic for determining the fixed network name
to use for test list server filters. Previously it had an unnecessary
if block to handle the case where getting a network name failed. This
simplifies the logic to just use a get() on the dictionary which will
default to false.

Change-Id: I356124459610342c8f7906397e23d58cbfad99bb