Check domain's enabled attribute strictly
Current test_domains checks domain's enabled attribute after converting
it to string because maybe that is for XML format. However, now Tempest
doesn't test XML format and the response of "create domain" is checked
as True instead of converting it to string.
This patch removes these conversions for checking boolean attribute.

Change-Id: I958de1e13b0169b03b56e4db1872d5ccf4e8904e