Revert "Test creation of server attached to created port"
This reverts commit 2b34ec3ff444e6e6ec7b3e52832ecd6e8ca20552

This breaks gate-grenade-dsvm-ironic-sideways on
stable/juno because the requested port's mac address is not
in the list of available mac addresses from the ironic
driver.

It's also unclear how useful this is given we already have
the test_preserve_preexisting_port test which is
essentially testing the same scenario, except it's not
run on stable/icehouse or stable/juno since preserving
pre-existing ports in nova wasn't fixed until Kilo.

Change-Id: I24403c1ae734b2137ddee5c3bf5a1594cf5375d8
Closes-Bug: #1445917