trim network smoke tests
This massively trims down the network api smoke tests to remove all
tests which require an optional extension from the tag, and pair down
what's left to a set of CRUD type tests for most of the major
resources. Additional tests for those resources which tested either
edge conditions, extra attributions, or field filtering are dropped
from the smoke tag as well.

Change-Id: Ia3472da9b0ef3103560ca23b5736dc6b492b1860