Make sure filtered_by_ip list servers test uses correct ip filter
The default tenants are now created with both ip4 and ip6 subnets. The order
in which they appear from a 'server show' is non-deterministic. This test
was just taking the first one and using 'ip' as the search parameter. For
an ipv6 subnet it needs to use 'ip6'. This test was only failing in the
non-isolated case because the tempest isolated tenant code uses only v4.

The ip6 case is now skipped due to nova bug 1450859 which causes the ip6
filter to always return all servers.

Change-Id: I242edb56bbfc9696d290c75c73918b93295efecf