Remove unnecessary force_tenant_isolation
The existing use of force_tenant_isolation seems to have been based on this
bug 1252971. We no longer have that
issue since there is no XML version, but I'm not sure why it was needed even
then if running tempest in a valid way. If tenant isolation is on then there
is no need. If not, then you are running serially, and there is no issue, or
you are using test accounts in which case there is also no issue.

Related-Bug: #1252971

Change-Id: I8bce43c73538128d530688f1e3c8ac9a0c119e68