Add cleanup after creating keypair
There is no cleanup added after a keypair is created in test case
test_create_specify_keypair in test_servers.py which causes the test
case to leave behind a keypair when allow_tenant_isolation is false.
This issue is not observed when allow_tenant_isolation is true and
admin credentials are provided in tempest.conf as the tenant itself is
deleted after the test case ends.

Closes-Bug:1453112
Change-Id: Ia17f4dd2585c0d51540b05bcd9e08eec2757c1dd
Signed-off-by: ahmad <afaheem88@gmail.com>