Provides a sample resourcefile for Javelin
Introducing a sample file in etc/ similar to the existing ones for
tempest.conf or accounts.yaml. This file can be used as a concrete
example of how to use Javelin.

Currently some resources are missing from the file (like volumes,
routers, and more importantly servers) and could be added in a later
commit.

Change-Id: I3e0d06c7654ad30c42e435a19106c6633144c9b5