Don't create network resources for scenario utils
The scenario utils classes make a couple of API calls to flush out
a matrix for using testscenarios on test cases with several different
images and and flavors. Previously, this was hard coded to only use
configured credentials, which may or may not exist in different
configurations. When it was updated to allow the use of configured
credentials no network resource dict was used which could result in
unnecessary network resources being created. This commit addresses this
by explicitly telling the cred provider class that no network resources
need to be created.

Change-Id: Ica3fa823771123a2c6e1d863f19a6006acb36f74
Closes-Bug: #1454490