Ensure scenario utils creds are cleaned up
This commit makes 2 changes to the scenario utils methods to ensure
that any created credentials are cleaned up after they are no longer
needed. The first change switches the ImageUtils class to not create
it's own creds, it is only called from inside of a test class which
has already allocated credentials. So instead this just passes them
into the class. The second is to add a cleanup helper on
InputScenarioUtils which gets called after all the api calls that
needs credentials are made.

Closes-Bug: #1455561
Change-Id: I07aa30091c8d0171b942e1804f5894d625363fdd