Add test caller to scenario manager ssh connection failure for tracking
test_volume_boot_pattern (v1 and v2) has been failing with generic
SSHTimeout for awhile now but due to the very generic nature of the
failure we don't have a fingerprint in elastic-recheck for tracking the
problem.

This adds the specific error to the message that's logged and adds the
test caller so we can track specific tests hitting this failure rather
than everything that could possibly hit it and muddle the bug
categorization in elastic-recheck.

Related-Bug: #1355573

Change-Id: I7b37af339aea1be0cbe4bae1156e9f5dc20d0dd7