Check for NotFound exception in setUp() function
test_resize_server_confirm_from_stopped test works fine when executed
alone but during a group execution it tries to delete the same instance
multiple times.

setUp() function is modified so that it catches NotFound exception and
create a new server instead of rebuilding (=deleting and creating) the
previous one.

Change-Id: Ie411268e37e4348a2434b473cd84e151a0a18c09
Closes-Bug: #1459516