Deleting ports before deletion of routers in resource_cleanup
In resource_cleanup need to delete ports before deleting the
routers. This is required in a private nuage thirdparty
testcase because a router-interface-delete is not allowed
when a port is present under that router. Also a port
resource sits beneath a router resource, and this change
would not affect any other tempest testcase.

Change-Id: I82a2cc9338a037f3b0d087c8f7c50c9cd0d0627b
Closes-Bug:1460223