Fix several bugs in verify_tempest_config
The version of get_versions() in tempest_lib somehow lost the removal
of the first argument so we need to capture it.
The version checking was looking for explicit minor version matches but
it seems that some services bumped them. This commit fixes the check to
only look for a match against the major veresion number.

Note that the unit tests were passing in spite of this bug. It would be better
if devstack failed when calling this script if there were any changes
suggested.

Change-Id: I0eba11d54b63dc2078e6987789933c1dc4650b71