Apply other "get" rule of GET to image v2 client
[GET /resources] methods should be "list_<resource name>s" or
"show_<resource name>", so this patch applies the part of the
rule to image client.

NOTE: This patch renames get_image_file to load_image_file
because the method gets the content of file. So show_image_file
didn't fit to its meaning. In addition, this renames store_image
to store_image_file for consistency.

Partially implements blueprint consistent-service-method-names

Change-Id: I2b354ff7192bea20339c61bedcd2f11eee455381