Fixing test_port_list_filter_by_ip in case of several DHCP agents
The problem is that the "test_port_list_filter_by_ip" test in
tempest/api/network/test_ports.py doesn't consider the case when 2 DHCP
agents (or more) service the network and this test may fail. So we may
encounter the situation when _fixed_ip_2 may be used by one of DHCP
agents. To avoid the issue we can create a port, not specifying a certain
IP address. The idea of the test is to list ports, using filtering by IP
and it doesn't matter what IP address from the subnet allocation pool
will be used. This approach allows us to avoid collisions between IP
addresses of ports and DHCP agents.

Closes-Bug: #1465676

Change-Id: I5e33d44c767a1e48f8fb28b03cab2f4627415f11