Added test_list_tenants non-admin test for v2, v3 api/identity
test_list_tenant test uses non-admin user's token and public
(non-admin) Identity API v2/v3 client to make a request to list
tenants as a user action. Test checks that user can see only
that tenants that he was bound to and cannot log in to tenant
that is not in his list.

Added 2 tests:
* test_list_tenants_returns_only_authorized_tenants to
  api/identity/v2/test_tokens.py
* test_list_projects_returns_only_authorized_projects to
  api/identity/v3/test_projects.py

Change-Id: Iabf3a474e70f87bc494d87d6333d668d52de2968