Provide a config option to customize remote shell command
Our remote ssh code makes all kinds of assumptions about the shell and
path of the guest. Recently some code was added to the command to generate
errors more promptly but it does not work if the shell is /bin/sh. This
commit allows the user to configure tempest with a desired prologue to handle
this case. It does nothing to solve the more general problem and, like the
other image-related options, cannot be customized individually for various
images that tempest may use.

Debug logging of remote commands was also added.

Change-Id: I669dd4386ffb539dad88a9487bef6c172b5d65fa
Closes-Bug: #1465682