update ceilometermiddleware sample target
current tests target storage.api.request meter. this meter is
being deprecated as part of the initiative to represent measurement
data as samples and non-measurement data as events[1]. this is to avoid
unnecessary calculations where volume is hardcoded to volume=1.
the correct meter derived from an upload notification from the
middleware should be storage.objects.incoming.bytes

[1] I654f657dd36967ad3ec99bbe06209e7745432e12

Change-Id: I8443ab9c1dc813bf0cc36f5195887f74e72f5859