Ensure server unlocking after test_lock_unlock_server
If assertRaises() in test_lock_unlock_server fails for some
reason, the server remains in locked state and cannot be
deleted after execution of test cases.

addCleanup call for server unlocking is added so that in case
of test_lock_unlock_server failure there will be no hanging
servers in the system after tempest run.

Change-Id: Iff336723d8e03814d9361492b1d5856b785f7a4c
Closes-Bug: #1468623