Re-factor neutron client for 'update' methods
Currently neutron client is written to auto generate
methods on fly. This is very difficult to debug and
make code very difficult to read.

As neutron client will be moved to lib, it will be better to
have its interfaces more readable and consistent with other
service clients.

This commit re-factor it for 'update_*' methods.

Change-Id: I67039beb934b88a048740a06e06d5a0a5b920942