Unbreak the world!
This commit contains 2 unrelated fixes, updates the sample config to
reflect changes in a recent oslo.log release and remove direct calls
to fixtures cleanUp(). Normally we don't want to do this and should
have only 1 fix per patch. However, because they are both breaking the
tempest gate at the same time and preventing anything from landing if
we don't land them in a single patch neither fix will be able to land.

The fixtures 1.3.0 release started enforcing that manually calling the
cleanUp() function for a fixture that was created using useFixture()
was a bad idea. It's basically saying we don't want to manage the
fixture lifecycle but at the same time want to manually run cleanUp().
This enforcement included in the new release has been causing several
unit test failures when running with that release. This commit
addresses this by removing all the uses of this call straightup as it
turns out it was unecessary, or changing the tests to not rely on it
anymore.

Change-Id: I24f6e20cc8e310ba69fb23510795e235218abb2d
Closes-Bug: #1469825