Removing unnecessary comments
There is a change (I24f6e20cc8e310ba69fb23510795e235218abb2d) that
removes code lines with direct calls to fixtures cleanUp() to fix
bug #1469825. But in this patch we forgot to remove some comments
related to removed code lines. So this patch is intended to remove
such comments. 

Change-Id: I693ae1dec027a9ea6752314f3f95c5b042dc6f95