Add tempest init command to tempest cli
This commit adds a new 'tempest init' command to the unified cli.
This command tempest init will be the first step of the new workflow
for running tempest. It is used to initilize a local working directory
to run tempest from. The idea being you can setup different local
working dirs for different tempest envs to test against several
different clouds.

The basic idea for this new workflow after installing tempest is
something like:

 1. mkdir tempest-newcloud
 2. cd tempest-newcloud
 3. tempest init
 4. edit local config file
 5. tempest run

The next step after this is to start working on the 'tempest run'
command. (which will include the hooks for the tempest plugin
interface) But, until that is added a test runner can be called
directly inside a local tempest working directory.

Partially Implements bp tempest-cli-improvements

Change-Id: I572267e61ba8ca86f92299e174864a7a89597207