Make the arguments of resource id consistent
Most arguments of resource id are "<resource name>_id" in compute
service clients, but some arguments are different.
This patch renames them to "<resource name>_id".

In addition this patch changes argument of fixed_ips_client also.
However, the argument name is not "<resource name>_id" because the
argument is ip address, not resource id.
The arguments of keypairs_client also are changed with this patch,
and new arguments are not "<resource name>_id" because a caller needs
to specify keypair name on these APIs instead of ID.

Partially implements blueprint consistent-service-method-names

Change-Id: I90cc456c0c49c79884bf0eb41fa93a9c602c1ef6