Remove wait_for_image_status from compute client
wait_for_image_status of compute client just calls the one of waiters
module and the method is a wrapper.
So the method seems redundant and it doesn't fit as service clients.

This patch removes wait_for_image_status from compute client and
replaces it with waiters' one.

Partially implements blueprint consistent-service-method-names

Change-Id: Ibce6fc48369c8c520c243926c4e0adf6da8eca5b