Remove wait_for_volume_status from compute client
wait_for_volume_status of compute client just calls the one of waiters
module and the method is a wrapper.
So the method seems redundant and it doesn't fit as service clients.

This patch removes wait_for_volume_status from compute client and
replaces it with waiters' one.

Partially implements blueprint consistent-service-method-names

Change-Id: I1e1bec6f7560efd59eb9962f49763316f1f02377