Remove wait_for_server_status from servers_client
wait_for_server_status of compute client just calls the one of waiters
module and the method is a wrapper.
So the method seems redundant and it doesn't fit as service clients.

This patch removes wait_for_server_status from compute client and
replaces it with waiters' one.

Partially implements blueprint consistent-service-method-names

Change-Id: Ie821ea4ba0924b71caba80132e1c383fb5bb125d