Switch all uses of json to oslo_serialization
This commit migrates all the usage of the json library to use jsonutils
from oslo_serialization. On python 3 httplib2 returns a bytes type for
the response body however all the methods in the json library are
expecting str types. We could switch all the uses of json.loads to
encode the input prior to calling json.loads however oslo_serialization
has baked in all dual python version edge conditions around this into
a call compatible lib which is much cleaner.

Change-Id: Icee30fb74db128c77b2c0c27e68b5801bd138cd5