Ensure a role is assigned to created users with v3 auth
This commit adds a default role assign to created users and projects
in the isolated credentials path if no other role will be assigned.
The issue previously was that when running with v3 auth tempest was
not assigning a role on the project for the newly created users. So
unless this was done out of band the created users would not be able
to get a token because they didn't have access to the project. By
always assigning a role to the user on the project this will no
longer be an issue. Additionally, we need to ensure that the default
role membership role we're using, "Member" exists before we try to use
it. If it does not it will be created before using it.

Change-Id: I4081cbd61f078bcc457062e2a55adb7b6f249a59
Closes-Bug: #1474193