Remove redundant attach/detach volume methods from test_minimum_basic
The nova_volume_attach() and nova_volume_detach() methods in
test_minimum_basic are also in the parent ScenarioTest class so just use
those rather than the local copies.

There is a similar duplicate code problem in test_stamp_pattern but that
test doesn't store all of it's variables as member variables so (1) it's
a bigger change and (2) test_stamp_pattern is currently skipped so there
is not a great way to test the changes without unskippping the test.

This is part of some work leading up to deprecating the
CONF.compute.volume_device_name option given nova change:

I76a7cfd995db6c04f7af48ff8c9acdd55750ed76

Change-Id: Ic8cfeb8358be480afb4f9d7aca879bd3abb5e775