Logging migrations for particular server if migration failed
Sometimes, test_live_block_migration fails on multinode job
(gate-tempest-dsvm-multinode-full).
http://logs.openstack.org/91/202391/5/check/gate-tempest-dsvm-multinode-full/f838b41/logs/testr_results.html.gz

If live migration is failed, server status become Active from Migrating but
host is not changed. In that case, test_live_block_migration fails for target
host does not match with server host. But from Tempest failure logs, we do not
know whether migration is failed or something else happened.

It will be better for debugging if this tests logs migrations information
for that server if live migration is failed.

Closes-Bug: 1476478

Change-Id: I0f7ba31e44e6d5508483a6552641f59828af20e1