Allow v3 identity to work without the admin domain name
The problem was that the value of admin_domain_name was required in order to
use identity v3 even if no admin was being used.

A new option auth.default_credentials_domain_name
is used instead of admin_domain_name except when requesting admin creds.
This defaults to 'Default' which is the name keystone uses for compatibility
with v2.
Because tenant_isolation and pre-provisioned credentials are mutually
exclusive, and to avoid having too many config options, the new option is
used instead of tenant_isolation_domain_name as well.

Change-Id: I52f0d4c0cc7e5eafa896776b12315ed6154dfae2