Add documentation to HACKING.rst describing idempotent_id
All Tempest tests require an idempotent_id decorator to uniquely
identify test functionlity. Previously this requirement was not
documented, leading to some confusion regarding its use and
implemtation. This patch includes documentation in the HACKING.rst
file that describes the purpose, format, and tool for working
with idempotent_id metadata.

Change-Id: I569bb68adc498f428664ed0b49d5943a353124b9