Full response for v1 ImageClient methods
Provide the entire response object, not just the images, for the
tempest v1 glance client's create_image, update_image and list_images
methods. This is all the work necessary to implement the blueprint
referenced below on this client as the remaining methods already
return the full response.

partially implements: blueprint method-return-value-and-move-service-clients-to-lib
Change-Id: Ia52bfde7cef1420c0c7d94288154fe4b6d54cd5f