Fix scenario test fails when port_vnic_type is set
The commit 0a03b7a37a20c206932206eea8602d9f4b009ae9 broke a working external CI
which tests SRIOV. This is a fix for this.

The main issue is that TestNetworkAdvancedServerOps object doesn't have
network attribute while TestNetworkBasicOps has. This patch set a common
attribute ports for all of them because ports are only created in
create_server method of manager.py when port_vnic_type is set.

Change-Id: Ib6c3081638bde05e56a40b70e2dda9677b43e04e