Fix volume backup import test
There is a bug in Cinder (#1455043) and backups are incorrectly imported
into the DB, and once that is fixed current tempest backup import/export
test will fail.

The reason for mentioned failure is that backup imports preserve
original backup id, so we cannot import a backup to a system where it
already exists, as is the case of our test.

This patch fixes this by importing the backup with a "fake" uuid so that
the import doesn't give an error.

Co-Authored-By: Yuriy Nesenenko <ynesenenko@mirantis.com>
Related-Bug: #1476416
Change-Id: I38e1a7b7d269278acd7963b990a263acd3d6e0d5