Full response from v2 ImageClient methods
Provide the entire response object for all methods of the
v2 ImageClient.  The equivalent change for the v1 ImageClient
(change 208045) included the temporary addition of a class method
that was to be removed when the equivalent changes were made to
the v2 client.  However, it turns out that for the method in
question the v2 client was already returning the full response
object which can be used as-is.  The additional class method
was therefore removed and a better change was made instead.

partially implements: blueprint method-return-value-and-move-service-clients-to-lib

Change-Id: I98c6a1685e4f3549faf1b567c6c6b193b6d644a2