scenario/test_minimum_basic: relax a MatchesDictExceptForKeys check
In the test_minimum_basic scenario, we check that a server that was
created a moment ago is the same as the server we just got. To do
that each fields of the expected vs actual servers are compared including
the "update" field. But the "update" could be updated asynchronously
by Nova, without Tempest knowing, causing a mismatchError.

This patch exclude the 'update' and 'OS-EXT-STS:power_state' (that
can be updated in background) fields from the comparison.

Change-Id: I50d1319b690453923b470733e94f3a11fd1cd249
Related-Bug: #1486475