Fix test_port_list_filter_by_ip with test_accounts
When running with test accounts test_port_list_filter_by_ip can fail
if the tenants network has overlapping ips. This is because the ports
list returned has more than 1 result because of potentially prexisting
ports. This commit refactors the test to take this into account and
check for the created resources in the list response.

Change-Id: I06bd027aa22103b65e7a081aa9cdfec5ebf7a360