add waiting for snapshot available status
In review https://review.openstack.org/#/c/213621/13
I've made a bug for some CI systems.

Code creates an image, but after, it doesn't wait that the volume
snapshot is in status "available". If the snapshot is in state "creating"
for let's say 30sec, the server created L200 will fail to boot with message:
InvalidInput: Invalid input received: Invalid snapshot:
 Originating snapshot status must be one of'available' values (HTTP 400)

Change-Id: I409eb77475ec3662f395fe0b52c5e03308c48973
Closes-Bug: #1489581