Add unit test for hypervisor_client
We have a plan to migrate service clients to tempest-lib from Tempest.
However, we are not doing enough tests for service clients now.
This patch adds unit test for hypervisor_client. After that, we can add
tests for each service client. The test identified an issue where the client
was using the standard json library to process the response instead of using
oslo_serialization; this is addressed in this change as well.

Change-Id: I401057ecdf70d693c285cdaf64e1a5db62fc9b17