Return complete response from remaining service client
As service clients are in direction to move to Tempest-lib, all
service clients should return Response without any truncation.
One good example is Resource pagination links which are lost with current
way of return value. Resource pagination links are present in parallel
(not inside) to top key of Response.

This patch makes remaining service clients to return complete Response body.

Implements: blueprint method-return-value-and-move-service-clients-to-lib

Change-Id: I4dd1315174fb102b165d565c2d4b9588d35cae50