De-duplicate test class names
Previously NetworksTestJSON was the name of 3 different test classes
in different modules. This was causing issues debugging failures in
one of them because certain class level events are only logged with
the class name to identify where the log message is coming from.
This commit addresses the issue by assigning 3 different names to
each of the test classes.

Change-Id: Ie4ffd458456d03b0b817b01bbed391f359240db2