Cleanup API version schema
There is "list API versions" schema for nova, and we will share the
part of it between "get API version" schema. Current schema contains
a lot of indents and it was difficult to make it with a single patch
for sharing. Maybe reviews will be also difficult. As the first step,
this patch separates the part of schema from "list API versions" one.
After that, "get API version" schema will reuse it later.

Change-Id: If711eaeb51ddca885a8df2fd3272ee8ea9ff2eaf