Add possibility to skip test_subnet_details if DHCP client not available
The help message of the dhcp_client option from the scenario section says
"...If left empty, update operation will be skipped. ...", but allowed
choises are only "udhcpc" and "dhclient". So we don't have any way to skip
the test_subnet_details test if DHCP client is not available. This patch
fixes the issue.

Change-Id: I10c403f4da257e6b39b855ae8b4b9ac16a8e7fc4
Closes-Bug: #1493909