Split the operator and reseller users in the account generator
The reseller admin and the operator roles have 2 very different
permission levels, and by combining them we break some assumptions
in the swift api tests about which users have permissions to perform
different actions. This commit splits the users created for swift
roles to be separate so users will either have the operator role or
the reseller admin role, but not both.

Change-Id: Ibd7db09095ca72bfe2ce231eb14a3d49637bf564