Remove _stop_instances hack in test_volume_boot_pattern
This was a carry over from when the test was originally written in
b3c3cb9c331e9ee54832326420ebe7ac4bf2e080 over 2 years ago.

We shouldn't have to stop the instances in order to delete them, that
would be a bug in nova if the delete failed for some reason because the
instances are still active.

Change-Id: I20b7c2e4c2b110933244ba65a1ccdbf246266946