Bring back Python 2.6 compat (don"t use dict comprehension)
We have a couple of places where we use dict comprehension. Dict
comprehension are just syntaxic sugar and can be replaced with
dict((k,v) for k,v in iterable) for Python 2.6 compat

I know Tempest doesn"t support Python 2.6 anymore and I agree it's
the good decision. But I am not a fan of "dropping support for the
sake of it", without compeling reason (like using a new module).

I am in charge of a 3rd party CI and my employer wants to have
Centos 6 support for OpenStack Juno. We use Tempest as our non-reg
test suite. Tempest is branchless.

This patch brings back Python 2.6 compatibility.

Change-Id: I23aeafcc9e68d09342a773d2e202f01884080fa3