Check cinder api version in compute tests
We've turned on Cinder V2 API as default so a few things
have changed.  One of those things is the response code
for volume/snapshot create.  These no longer return a
200, but instead a 202.

The result is that some tests are hard coded to still
use the default cinder client API version.  This
means that those tests will receive the wrong response
code.

This patch just adds a check to common/compute.py
where we setup for the volume backed migration test.

There is probably a better solution to do this further
up and propogate it everywhere, but this provides a
quick fix for the one know case that fails.

Change-Id: I3c1524ade051a8ace1e393085efb1d24fc832c9b
Closes-Bug: #1500923