Fix return value of reserve_fixed_ip client method
In Fixed ip service client, reserve_fixed_ip method does not
return the response body.

As per consistency in service client interfaces, this method should
return the response value even that is empty body.

This commit makes reserve_fixed_ip method to return the resp body.
Also implement unit tests for the same.

After changes are merged in tempest these needs to be sync in lib.

partially implements: blueprint method-return-value-and-move-service-clients-to-lib

Change-Id: I2cb1ef8dacd1994576656cfb108dd2eef2fbd988