Remove preprov provider dependencies from CONF
This commit removes the last dependencie of the preprov
provider from CONF. The required configuration is passed
now as parameters at init time.

As the number of parameter grows, and many of them are taken from
configuration, adding dicts with parameters that can be used
when instantiating the credentials providers.

Depends-on: I0ca0b96fc618ffe60851984a9c6d46b0507878d4
Partially-implements: bp tempest-library
Change-Id: I945994580ce9a29f6fbf67183e5da4b100542386