Pass the status 'ACTIVE' for getting a target port
The gate test failed because Tempest could not distinguish a target
port for floating-ip due to existence of multiple ports. The one port
was the status 'ACTIVE', but the other was 'DOWN'.
At least a target port should be ACTIVE, so this patch makes Tempest
pass the status 'ACTIVE' for getting a target port.

Change-Id: If5b9054284547099d8a64a2e3a608e9810b22067
Related-Bug: #1505947