skip compute alt tenant tests if no project_id in url
A set of the alt tenant tests in Tempest assume that you can route a
request to an alternate tenant by changing the service url. This is
only true if the project_id is part of the service url. We're moving
to a model where that will be optional in Nova, in which case these
tests don't test what they think they are doing.

Put in a simple check that if client.base_url == alt_client.base_url
then these tests won't work as expected, and skip them.

Redo comments as documentation for the tests touched.

Change-Id: I12d41255e7b809111fd26eb6588ea24e9b4cb3af