Fix DEFAULT_CONFIG_DIR
The current DEFAULT_CONFIG_DIR makes no sense
when installing tempest as a python package as it is
derived from the absolute path of config.py which is
incorrect.
For example:
/usr/lib/python2.7/site-packages/tempest/etc/
Is wrong.

Instead, it should be set to the relative path from
where the command is being executed.
For example, if tempest init ~/my_tempest were run,
the correct default config directory would be:
/home/me/my_tempest/etc
for the “my_tempest” initialized runtime.

Change-Id: I57d62107182ce652bbfb3c077b8a2311393234ed