Only run dual-NIC/dual-stack IPv6 tests in gate
This reduces test time by 75%, which in this case is many
minutes.  The speed-up in the gate queue is significant.

We don't really need to test the other six modes I'm marking
"slow" here, as Dual-NIC/dual-stack tests everything, and if
it's not working then the single-NIC case doesn't matter.

They will still get run on a periodic basis everynight, with
status available here:

    http://status.openstack.org/openstack-health/#/job/periodic-tempest-dsvm-all-master

These should eventually move into Neutron as in-tree tests.

Closes-bug: #1489962
Change-Id: I378710382fb01c5afaf10c5e892600244be02449