make schedule_to_every_node use real scheduler hint
force_hosts is a scheduler construct, however it's not one you can
directly set via the API. Because we don't enforce schema on hints it
was silently ignored. The supported way to do this is through a
slightly baroque pass through in availability zones.

availability_zone: "$ZONE:$HOST"

This can only be done by admins by default, so we need to handle to
also replace self.server_client with self.admin_manager.server_client.

Change-Id: I3e61f039acf3278e0231183c4a774caa91681f7a