test_port_security_macspoofing_port: Fix an inversed check
As create_floating_ip() associates the fip to a server,
the status of the fip will get ACTIVE sooner or later.

For the Neutron reference implementation, the current test
which ensures the fip DOWN succeeds only when it happens
to run faster than L3 agent.  In case of MidoNet, the test
always fail as the fip gets ACTIVE synchronously.

Closes-Bug: #1521185
Change-Id: Ie6d4f5428dccdf4c83a647361cdae0c739fae394