Update image creation error msg to mention the correct image type
Currently, the glance_image_create method allows the ability to
create any type of image as specified in the [scenario] section
of tempest.conf. The error message when unable to find this image,
however, only mentions qcow, which can cause confusion for users
testing with other image types. This commit updates the error to
mention the image type configured in the conf and also issues the
log as a warning rather than a debug message.

Change-Id: Ia2e81da2ded351df5b53f36de278a3400d1b2605