Add compute client base class for microversion support
This commit adds the base class for compute service clients
to pass the microverion in request header.

Microversion is being set on service client at class level
which will be used to send in microversion header of each
method of that service client.

Further each compute service clients will be derived from this
class to support microversion testing.

We are currently doing it in Tempest and after getting enough
proof of testing, this will be sync to Tempest-lib.

Partially implements blueprint api-microversions-testing-support

Change-Id: Ib7bb5804a0b6e55d189efc742e61412cb42d7e79