Remove pre-check for test accounts file in credential factory
This commit removes a check for a valid accounts.yaml file before
we use the preprov cred provider. This masks real configuration issues
and makes it very confusing for users to see when something doesn't
work because it defaults to trying to use the legacy provider. Since
we'll actually fail to load a missing accounts.yaml file now this also
adds a check for IOError and raises an InvalidConfiguration exception
with a descriptive message instead.

Change-Id: I9c02b406746af799f8939e0b9d06c14b48302a7e