Adds default domain name to dashboard login params
Desc:

The current tempest configs requires a new property for the version of
Keystone that Horizon is set too. The login screen will require a
domain name when Horizon's local_settings.py is set to Keystone v3 with
multi-domain support enabled.

Details:

The Keystone API runs both version 2 and version 3 concurrently.
They are not compatible and have different methods of authentication.

Clients can choose which version of the API to talk to. (i.e. choosing
the endpoint)

Horizon (The OpenStack Dashboard) uses the Keystone Client binding to
authenticate to the Keystone API. Too choose which version of the
Keystone API to use there are settings in Horizon.

When Horizon's settings are configured to talk to the Keystone V3 API
the login screen will require additional credentials. Namely the the
domain name of the user. When this is omitted, authentication will fail.
The user will not be able to login and view the Overview screen.

Tempest has a scenario test that assumes Horizon will be configured to
use Keystone V2. This tests fails when Horizon is configured to use
Keystone V3.

The Proposed Fix:

Pass an additional request param to the login form that will allow
this scenario test to work for both keystone v2 and keystone v3.

Closes-Bug: #1531594
Change-Id: I060a7ceb19c2c2872b065c94d31255b0dc1750fc