Mark Javelin as deprecated
This commit starts the deprecation timer on the javelin utility and
adds a deprecation on stdout when the command is invoked. We stopped
using it in the upstream gate when grenade was refactored as part of
adding the plugin interface. We'll remove it from the tempest tree
after a standard deprecation window has elapsed.

Change-Id: Ie4ab5258946271a16e96a070abeedfafa5423714