Mark amazon image file config options deprecated
This commit marks the 4 options used in the scenario section as
deprecated. The behavior of these options and the other image file
options in the scenario section have a very convoluted and difficult
to understand. There is no reason for this complexity we should just
have a single option to tell tempest where the image file is. As the
first step in cleaning this up this marks the options which will be
removed as deprecated to signal to users the future change here.

Partial Bug: #1393881

Change-Id: I8a6ee167384c8a2d17905db3a4061c06398980ad