skip test_list_server_filters_by_regex
The test for the regex only works if we are always allocating linearly
from the bottom of the network block or on a small network
block. We're changing the IP allocator in Nova to be more resiliant,
and this test can be proven to be trivially wrong in that case.

It should be skipped until it can be fixed.

Change-Id: I27f73c1edf12218818c4d279efbd9fef5cdef672
Related-Bug: #1540645