test_images_negative: Fix URL of invalid server
The requested server ID contains the hash tag, which makes the sent request url suffix of the form: `/v2/tenant-id/servers/!@#$%^&*()/action`, which includes a hash tag. The path of this url is `/v2/tenant-id/servers/!@` and the fragment identifier of this path is `$%^&*()/action`. This is not the right behavior and might lead to failures when a backend server parses the url correctly.
Removing the hash tag from the ID will fix this.