Fixed bug with url building
The requested server ID contains the hash tag, which makes the sent
request url suffix of the form:
`/v2/tenant-id/servers/!@#$%^&*()/action`, which includes a '#'.

The path of this url is `/v2/tenant-id/servers/!@` and the fragment
identifier of this path is `$%^&*()/action`. This is not the right
behavior and might lead to failures when a backend server parses the
url correctly.

Removing the '#' from the ID fixes this.

Change-Id: I1ce02e78324ccbca314218a8e62967db68e59a80
Closes-bug: #1541761