Allow user to be disabled
When a user is disabled users_clients.list_users() is called.
In the above method it is assumed that they are looking in the domain
that is specified in the tempest.conf.  In test_users_negative.py. The
user are created in the default domain.  To fix this we need to passed
the domain_id to list_user to get the user in the correct domain.

Change-Id: I348600da7c7c12e0dcbfcf6f4e376e2950b31a5f
Closes-Bug: #1549044