Properly handle failures during resource cleanup
The current code that cleans-up network resouces from the dynamic
credentials provider does not account for failures during the
cleanup itself (apart from 404).

When the cleanup is triggered by a failure during the setup, the
original failure is hidden which makes debugging impossible.

Catching all exceptions on cleanup, logging a warning and
re-rasing the original exception should fix that.

Fixes bug 1551264

Change-Id: I49dbb314a09fd9b49c9604bc8b5b8ed03ff218d4