Add LOG.info to boot from volume tests
These tests fail a lot, more than nearly any other tests we have. When
reading the REST logs it's a bit hard to figure out what we were
conceptually doing because we repeat many of the same actions over and
over again.

This adds some LOG.info to provide a road map when looking at logs.

Change-Id: Ic9e3bf9b5720d6f250c9b003357ba4bc5fdebbe3