Make data_processing/baremetal use rest_client
Now service_client is unnecessary because that is the same as
rest_client. This patch makes both data_processing_client and
baremetal_client use the rest_client, and we will remove
service_client for the cleanup in the future.

This patch removes these clients from a certain test because
this test is just for service_client, not rest_client as the
test name. We have added this test for verifying all arguments
are passed into rest_client via service_client. However, this
patch some clients use rest_client directly, so this test becomes
meaningless for these clients. That is a reason why the migrated
clients(into tempest.lib before) are not contained in this test.

Partially implements blueprint consistent-service-method-names

Change-Id: Icf85bcde4c0d1ee644c950323f4c6abe3b86a24e