Make database/telemetry clients use rest_client
Now service_client is unnecessary because that is the same as
rest_client. This patch makes database and telemetry clients
use the rest_client, and we will remove service_client for the
cleanup in the future.

This patch removes these clients from a certain test because
this test is just for service_client, not rest_client as the
test name. We have added this test for verifying all arguments
are passed into rest_client via service_client. However, this
patch some clients use rest_client directly, so this test becomes
meaningless for these clients. That is a reason why the migrated
clients(into tempest.lib before) are not contained in this test.

NOTE: This patch changes __init__() of versions_client to use
      **kwargs because service_client removed them if each value
      is None, but rest_client doesn't do it. The purpose of the
      __init__() is just for calling self.skip_path(), so this
      change is more readable also.

Partially implements blueprint consistent-service-method-names

Change-Id: Ie99bbbe7655138c4c3ecae69ae0b361198a9651d