Make identity clients use rest_client
Now service_client is unnecessary because that is the same as
rest_client. This patch makes identity clients use the rest_client,
and we will remove service_client for the cleanup in the future.

This patch removes these clients from a certain test because this
test is just for service_client, not rest_client as the test name.
We have added this test for verifying all arguments are passed into
rest_client via service_client. However, this patch some clients use
rest_client directly, so this test becomes meaningless for these
clients. That is a reason why the migrated clients(into tempest.lib
before) are not contained in this test.

Partially implements blueprint consistent-service-method-names

Change-Id: I5de906bcc0e32063ff17ee75f8dcf91b99165d03