Implement TIMEOUT_SCALING_FACTOR for tempest
This is a pattern we've had in Nova for a while that allows individual
test classes to be marked as slower. This timeout adjustment is needed
for test_volume_boot_pattern because going up and down with 3 guests,
and cleaning them all up, plus volumes and snapshots, takes
time. Especially on slower personal equipment. These fail for people a
lot, which makes them think their devstack is broken. And it means
that it leaves crud around when it times out, which ruins follow on
runs.

Change-Id: Ibf21bdf44339d00f0291d0ff14eb6bcc69798ccc