network: Fix subnet_allocation extension check.
This code checked for a "subnetpools" extension.  While it's commonly
referred to that way, and may have originally been called that, the
API extension is actually "subnet_allocation".  Nobody has probably
noticed this error as the default neutron jobs run with the
api_extensions configuration set to "all", so all checks like this
would return True.

Change-Id: I646b5e988bc6a9053d5361d81ed4c166fc670c7a
Signed-off-by: Russell Bryant <rbryant@redhat.com>