include domain_id when creating groups
as of the liberty release, keystone has emitted deprecation warnings
when groups were created without a domain id specified; keystone would
fall back to using the default domain specified in keystone.conf.

as of newton, this fallback is being removed in favor of always
specifying the domain id, and raising an exception if it is not present.

Change-Id: I6004644531c2e86cafe3714451101e6b977a8593
Needed-By: I0b061c7df45c10b419c0b1474c9217c4402f70fb