Removing wrapper method for "wait_for_volume_status" function
The wait_for_volume_status method in the base_volumes_client.py file is
a wrapper method that just uses the waiters.wait_for_volume_status function.
The wait_for_volume_status method is not a service method and that is why
it should be removed and the waiters.wait_for_volume_status function should
be used instead.

Partially implements blueprint consistent-service-method-names

Change-Id: I4516a2cc2dd6ada8d0a49392a98b949edc09a9fd