Use common "waiters.wait_for_snapshot_status" function everywhere
In Tempest we have the waiters.wait_for_snapshot_status function that
is already used in some places. So this commit replaces the use of the
wait_for_snapshot_status method defined in the base_snapshots_client.py
file with the use of the waiters.wait_for_snapshot_status function.

Change-Id: Ic017d468cc1478d8207ba482f161ed63c9b168fe