Snapshot list using 'limit' parameter
This patch tests the option to determine the number
of snapshots that will be displayed from a list of snapshots,
by using the 'limit' parameter.

The patch caught 2 cinder bugs when limit=0. One in api V1
and second on kilo V2. In both cases api returns,
all snapshots, while we expect to get an empty list.

Change-Id: I1b83d98ec0d8e97adc849a3fdd1471f3ffc5cbb7