Updated Image Not Created Error Thrown
When image creation fails in this instance, we get a
delete image error. The error is occuring where the image isn't even
created in the first place. This patch makes it so the error will be
that the image is never created, instead of failure of deletion.
It will make the stacktrace easier to understand in the case that
someone runs into this error in the future.

Change-Id: I2f38685d36855fe23dea6624a9c9aef04a5158f9