Addresses Expect: 100-continue client behavior
Updates to the object storage object services cases with
Expect: 100-continue. The current test case sends data
along with the Expect: 100-continue header, which does
not properly mimic expected behavior. Updated the
method to send no body with the proper content length
along with the Expect header then sending the body once
the 100-continue is received.

Closes-Bug: #1573859
Change-Id: I91b486e067a0acacf7ca121e6d5da006554b5348