Added logic to validate storage policy info
Added storage of Swift version and storage policies for future tests to
reference.  Added logic in the custom matchers to leverage that data to
evaluate if storage policy usage is in the account responses.

Added logic in the custom matchers to also check the x-storage-policy in
the container response against the stored list of policies.

Change-Id: I8496036274117132bae6146353b778577286ee65