Add option to tempest init to show global conf dir
Due to the fun and nondeterminism that is python packaging knowing
where tempest thinks the global state dir which is used to preseed
the local etc/ directory can be a complete mystery. To address this
issue this commit adds a new option to tempest init that will print
where this directory is and then exit. So in the off case you want
to preseed your local tempest working directories you can use this
to figure out where to put the files.

Change-Id: I07128fb802a6c1408d4f6154c68004b38aaf1860