Change scenario test 'test_resize_server_confirm'
test_server_advanced_ops.test_resize_server_confirm is a duplicate
of test_server_actions.test_resize_server_confirm, so rather than
delete it for being redundant, let's enhance it to be different.

There are currently no resize/migrate tests in Tempest which
resize a volume-backed instance, so this changes the scenario
test to use a volume-backed instance for the resize.

This change is pretty simple, but anyone that was blacklisting
the test based on the name would be broken. However, they should
be blacklisting on the test uuid anyway, so meh.

Related to nova blueprint libvirt-instance-storage

Change-Id: Ie91b2881cb59638ac6e533ab93024dbc42111772