Modify projects_client to receive more attributes
Starting in Keystone v3.4 the parent_id attribute was introduced for
projects as a way for creating projects within a hierarchy. Also in
v3.6 the is_domain attribute was added. The pojects_client that is
part of the identity v3 clients implemented in Tempest was missing
these attributes. Instead of adding the individual prameters to the
client, this commits allow users to pass as many parameters as they
want to the create_project and update_project methods through the
kwargs, so whenever new attributes are added to Keystone they can
be used in the identity v3 client in Tempest.

Change-Id: I171fa21113298c1e01ddc98ebd1791c46223d69b