Add assertion to test_create_security_group_rule_with_invalid_ports
This patch syncs the test from Neutron to Tempest so we can delete
it from the Neutron repo. The API change and test addition was
originally authored in patch with Git hash prefix 91c476dc.
The API change is present from Liberty forward so should be
safe to add to Tempest since it's no longer testing Kilo.

Change-Id: Ic77561757b5bd9ac61b1e059afdba19a98257126
Related-Bug: #1552960
Co-Authored-By: huangpengtao <huangpengtao@huawei.com>