Unit test: fix flaky test_networks_returned_with_creds
Test test_networks_returned_with_creds is run twice, once for Credentials V2
and once for Credentials V3. They shouldn't run at the same time because
they have a side effect to create/or read the same directory. (look at
`_get_free_hash` in preprov_creds.py).

We fix this race by letting each test own a different directory.

Note: the proper fix is to properly mock the environment. But this is
beyond my understanding of the code. This fix is good enough imo and
fixes a Gate failure.

Change-Id: Ia324c510e91ede4fb387c310ad2a24d15325058e
Fixes-Bug: #1586640