Deprecate run_tests.sh
This commit deprecates run_tests.sh script and adds a deprecation
document. We don't pay attention this script actually however, I suppose
new people are coming and see the top of the directory and find this
script and try to use this. And if something wrong happens, we need to
ask and check 'which runner did you use?'. So I would like to reduce the
ways to execute unit tests.

Change-Id: Ia764fa3a15dc9faa753d4aca4c76462a8b9dd40b