Increase test coverage on preprov creds
Add a few more tests for the preprovisioned credential provider,
for the get_*_creds methods that were not explicitly covered until
now. Fixing a small issue that was detected by the new tests, when
no credentials of one type are defined in the YAML file, the wrong
exception was raised.

Change-Id: I5727baf95f50e55aff17232862c2b0705f1eb743