Check which glance client version to use
Nova is working toward supporting the glance v2 API which will
allow disabling the glance v1 API in glance in a deployment and
configuring Nova to only use the glance v2 API.

The existing API and scenario tests are setting up a base test image
using the glance v1 client regardless of whether or not the glance v1
API is enabled in tempest.conf.

This adds a check to the tests to see which glance client version
to use depending on the config.

Change-Id: I1b1dc2911698d4e4d1a2727902ba457688d88752
Closes-Bug: #1589093