Make missing global config dir not fatal in tempest init
Due to the complexities and inability of the python packaging
infrastructure to install data files properly or consistently we have
to be prepared for the case when we have no idea where the data files
end up from a tempest install. Thus this commit makes tempest init
not fail when it can't find the global config dir and just logs a
warning that it couldn't be found.

Change-Id: Iff944092228ca1ae081e8faaf10ee0048e7ea414