Factor up (most) CONF value in clients.Manager
Factor up CONF dependencies from the client manager class, so
that the part of it independent from CONF and unstable interfaces
can be moved to the manager base class, and be useful to external
consumers (such as plugins).

An helper function is defined in config.py to help extract from
configuration the parameters relevant for service client __init__.
This is used for Tempest own service clients and will be used by
plugins as well.

Partially-implements: bp client-manager-refactor

Change-Id: I19cdcf6f84144f6b1b0223a41a52f5b6e4b8fdd1