Adds clone/snapshot volume test for create volume
Cinder supports creation of volume from snapshot, image, or another volume as
source. In the case of snapshots and volume sources, its allowed that the
created volumes are bigger than the source. Cinder has had several backends
that has this checking broken. We need to have this tempest test running in
Cinder CIs to identify possible bugs and test submitted fixes that meant to fix
those bugs.

Co-Authored-By: Marcus V R Nascimento <marcusvrn@gmail.com>

Change-Id: I829c0e6f99fff0ec0bd5366c9d2e79b1b073f0a3