Add test removal procedure doc
This commit copies the Tempest test removal procedure doc from wiki[1]
to the tempest repo. Because this document is tightly coupled with
tempest so we don't need to manage it in the wiki. And we can see and
review this through the gerrit when we would like to update it.
And I'll update the wiki page to be just a link after merging this.

[1] https://wiki.openstack.org/wiki/QA/Tempest-test-removal

Change-Id: I5f2322a5de7b73252e0162165eeaec120ec94948