Remove the _try_wrapper function from identity tests
The _try_wrapper() function was used on the cleanup path for several
identity api tests to wrap deletes and ignore exceptions. But, this
was a terrible thing to and just masked real bugs. If an api call
fails for an unexpected reason, even in the teardown path, that is
a real problem. This commit removes this and replaces it with just
ignoring 404s which is a possible expected condition on teardown.

Change-Id: I6996a8320c510923172a157d3257710c1e9baaa3