Member role may already exist
The tests that first setup the v3 Member role may be getting into a race
condition where the Member role is attempted to be created by multiple
workers. This change ignores the conflict because the Member role is not
test specific so it can only be created once when no domain id is
provided.  The create_user_role call already checks if the role exists,
but we have seen multiple threads attempt to create the role almost
simultaneously leading to a conflict on the second request.

Closes-Bug: #1596250
Co-Authored-By: Alex Schultz <aschultz@mirantis.com>
Co-Authored-By: Matthew Treinish <mtreinish@kortar.org>
Change-Id: Ic96f622a2fa00e2fd0cb3ebb22e3df000bac5345