Fix minor details in README
This commit makes 3 updates to the README. First it adds a link to the
official documentation at the top of the README. We've had a number of
bugs filed where people think the docs are broken because github
doesn't understand sphinx directives in its rst rendering. Including
a link to the openstack hosted logs will make it clear that the docs
output is only via sphinx. The second cleanup is adding an internal
link to the tempest configuration guide in the quickstart. The
quickstart alludes to configuring tempest, but never actually provides
an explanation on how to do it. Adding the link to the configuration
guide should clear up any potential confusion. The last change here is
removing the external shield images from the readme. These raise sphinx
warnings (which we want to stop), and violate at least one distros
packaging process for questionable value.

Closes-Bug: #1589428

Change-Id: I0f797478624df704c03a00bc3fae89bfcea5799a