Check network_data info in config drive tests
There is an existing scenario test that verifies the meta_data.json
contents of config drives. Extend this to also do verification of the
network_data.json information on the config drive.

The network_data has been included in various ways by different cloud
deployments so we want to have a concrete test for this somewhere as the
first step towards consistency.

Change-Id: Ia148bca2cd932ccdf9735142e0082a330212ee04