Don't enable DHCP on floating IP subnet tests
The test 'test_create_update_floatingip_with_port_multiple_ip_address'
creates a port with multiple fixed IPs by retrieving unused IP addresses
and then assigning them.

Change I84195b0eb63b7ca6a4e00becbe09e579ff8b718e adjusted the unused IP
selection logic so it now picks IPs close to the start of the CIDR so there
is a potential race with the IP for the DHCP port. (the previous logic chose
IPs at the end to avoid this)

This patch just disables DHCP on the subnet since it's not needed for the
tests to avoid racing the DHCP agent.

Change-Id: I17e201ef8822cace86bf805d6bd5a2e4d0e9084d
Closes-Bug: #1600398