boot into a network that has "port_security_enabled=False"
If the tenant has a Neutron network with disabled port security and
no security group gets specified when booting an instance, the ports
attached to the instance should not be associated to the 'default'
security group.

NOTE: Commit ee7a0198261 documented this known issue with a release note
"known-issue-on-api-1efca45440136f3e.yaml" and cleaned the "default"
security group from the Nova API.

This adds a feature toggle config option for making the test run where
the bug is fixed, which is mitaka and newton. This allows us to run
the test in mitaka and newton but not on liberty where it still fails
and we won't be backporting the fix. There will be devstack patches to
enable the test in newton and mitaka.

Depends-On: I630008a9733624a9d9b59b7aa3b8b2a3f8985d61

Co-Authored-By: Matt Riedemann <mriedem@us.ibm.com>

Change-Id: I20b8d5d2a300c83a59bdb33374fc20447ce2ede3
Related-bug: 1175464