Add a test for attach/detach port on multiple servers
A regression was introduced in nova in newton when you
attach the same port across multiple instances (not at
the same time). The problem was nova started creating
some internal resources on attach but wasn't cleaning
them up on detach, so when you'd tried to attach the same
port to a second server it would fail because of a
unique constraint on the internal resource that nova creates.

We should have an integration test that covers this scenario
so we don't regress it again.

This change covers both booting servers with a pre-created
port and attaching a pre-created port to existing servers.

Depends-On: I2254bad0df3ccc00cd5c9438fa2684e705442e2d

Change-Id: I469b8ec426bd71dea515e99f76d415c62fff7dd3
Related-Bug: #1602357