Remove DataGenerator from admin V3 identity tests
The identity tests have a DataGenerator that handles resource creation
and cleanup for tests, however using this DataGenerator has proven to
make debugging more difficult. The object storage tests used to use
this DataGenerator as well but was removed for the same reason. This
is the second of a series of commits to removes the DataGenerator from
the Identity tests. This commit removes the DataGenerator from the
admin v3 tests and moves resource creation and cleanup back to the
tests or the resource_setup, resource_cleanup methods.

Change-Id: Ic73cbffec8ff801462ea2155718fc3728bfb6512