Add tempest test to test NoVNC support
Currently there exists support within tempest to make sure that the
VNC Console URL can be generated, but there is no support to test out
the actual NoVNC code through the novncproxy.  This will add support
to verify that the novncproxy will return the proper NoVNC JavaScript
HTML, that the authentication token is validated, and that the basic
exchange of the RFB protocol from the client through the novncproxy
is a valid VNC session.

This test requires that the tempest configuration specifies that the
VNC Console feature is enabled, otherwise the test will be skipped.
Additional integration test environments can be switched to enable
this feature to take advantage of this test.

This support is being added in conjunction with change set 115484 to
further test the NoVNC TLS Support, but this test case will work
whether or not the communication between the novncproxy and VNC
session is using TLS or is unencrypted.

Change-Id: Ic7e149f79705cbb910c5b6524dac5cdb73d69710
Closes-Bug: 1554460
Depends-On: Idb38a3b11e2f61f23adf1ec23c04ddccd72e7539