Fix release notes around
We need to add new release notes files added for each commit. While we
can edit past release notes that's essentially the same as editing
history, it's fine for typos and/or clearing something up but when
adding new things we're adding onto notes in past releases. In the
case of 2 recently migrated clients the old release notes putting
patches that landed after 12.1.0 in the release notes for 12.1.0. This
commit fixes that issue and breaks that out into a new note.

Change-Id: Id7a24c1bc659c4e44b2214ec27e6b94ea742a81e