Add compute API tests for 'get-me-a-network'
This adds the tests for the 2.37 microversion in nova
that handles automatic allocation of network resources
for a server via neutron's auto-allocated-topology extension
when there are no networks already available to the tenant.

This also adds the 2.26 microversion json response schema
since that's what's coming back for GET and LIST operations
on servers when testing the 2.37 microversion.

Tests for nova blueprint get-me-a-network

Change-Id: Ia8538e20ba52c85cfabbf9e44a959b2821995eab