Switch tox to use tempest run and deprecate bash runners
This commit switches the tox commands that run tempest to use tempest
run instead of pretty_tox.sh. Tempest run is the primary method of
running tempest moving forward so we should dogfood it where we can.
This also officially starts the deprecation timer (which was long
overdue as these were unofficially deprecated before) on the
run_tempest.sh and pretty_tox scripts as we move to having a unified
method of running things.

Partially implements bp tempest-run-cmd
Depends-On: I8786b9709584378aaf48393115dda94397e22544
Depends-On: I0529c2922242d8ed9c92ef718328e09f3ea9e05b
Depends-On: If0064b9e8358332972ef4a1eee8f150e66f8c50f
Depends-On: Ibaffa3a8568ea058d964463df4b76196c4d2bc7a
Change-Id: I3684fce66a799579fa68af119652cafef25a9f03