Use expect_empty_body flag
The create_router_on_l3_agent and add_dhcp_agent_to_network API
expects to get a response without the body so their tests must
use expect_empty_body flag.

This patch also makes the body to be None because now "null" is passed
as body. Empty string will be passed once the issue is fixed[1] and
then the setting will be not needed.

[1]: https://review.openstack.org/#/c/291545/

Change-Id: If33f518b817f983a107a736172026d86c47f9878
Related-Bug: #1555921