Revert "Move dscv and ca_certs to config section service_clients"
This was a backwards incompatible change that breaks tempest's stability
guarantee on tempest.config. While the option was marked as deprecated,
oslo.config does not alias deprecations in code. The fact this patch
had to update unit tests was an indication of this. This breaks anyone
who consumes these options via code. We can only land something like
this after we add aliasing to oslo.config for deprecations.

This reverts commit 1afca56b059e07bfb7ae360fe0c3a753719aea92.

Change-Id: I91600d8fc49db670b0f78ffd0094575fbfa0ca63