Separate encryption_types_client from types_client
The is_resource_deleted() of types_client covers both volume_type and
encryption_type and the method interface became different from the
other service client, because the existence check way is different
from both resources. That is not so good for stable interface and
it is nice to separate encryption_type way to simple code.

Partially implements blueprint consistent-service-method-names

Change-Id: If4ecd61d2886d616b677d8bbeaf31646cb1b19cd