Add port parameter to ssh Client
This commit adds a port parameter to the ssh.Client. We don't use it in
our test cases. However, it would be good to have it when I debug this
class because my development host ssh port is not 22. And I think
that it is natural a ssh client has this feature.

Change-Id: Id8b73108f67fe8c581dd8fba90902869b4ee776d