Add existing volumes when using pre-provisioned credentials
When using pre-provisioned credentials, the tenant may have
volumes other than those created in the testcase and so will
make test_volume_list_param_tenant fail unexpectedly. e.g. the
existing volumes may be those not being deleted successfully
and ending up with status 'ERROR', which is not rare seen in
production clouds. So this is to add existing volumes in
test_volume_list_param_tenant.

Change-Id: Ice975b2011e7914d2b92ea71e5dd30f9074f6243
Closes-Bug: #1618730