Narrow assertion in attach/delete volume test
The attach_delete_volume test takes the device name
from tempest.conf's volume_device_name and looks for the output
in /proc/partition.  It could be that this name occurs in part of
another partition's string.  Patch adjusts the name used in the
assertion to make sure it's an exact match.

Change-Id: I570864c7c38969dcdcbfa5bd71be0348d90bd790
Closes-Bug: #1619564