Don't create network resources for identity tests
The identity tests used to create network resources but they are not
needed because we are testing the Identity service. This patch
stops creating these network resources.

Note: somehow security groups are still created and deleted. One can
look Neutron server's log while running the identiy tests to see this.

Change-Id: Ie77d8ecff5f346ed9411192d7076c321a9a22929
Closes-Bug: #1618519