Use local variable for volume instead of instance one
The patch is a follow up to finish replacing instance variables
by local ones for volume, as it has been done in [1] for server.
Also, the _delete_volume() helper function has been replaced by
self.delete_volume method from base class, which waits for volume
to be gone itself, so we don't need to use an extra check for it.

[1] https://review.openstack.org/#/c/333305

Change-Id: Id217d7a374c663982a9f97b83c72923136305f9f