Always wait on server delete
When running with neutron not waiting on the async server delete can
cause the port cleanup to race with tempest's other network resource
cleanup and causes things to explode. This commit makes adding the
cleanup wait loop unconditional there is no reason this needs to be
configurable and tempest should ensure everything is clean before
it starts tearing down everything else.

Change-Id: Ica2f9bb20a059ab35577878c41172277e87df8e6
Closes-Bug: #1623691