Add config option to create networks with port_security_enabled
The port_security extension in neutron is disabled by default.

devstack enables the port_security extension for ML2 but since
it's optional we shouldn't assume it's enabled, so this change
adds a configuration option to tell us if port security is
supported as a networking feature.

We could use the extension list but that defaults to all and
we shouldn't require tempest users to have to maintain a
whitelist of extensions just to make the default configuration
work with default neutron, i.e. not devstack.

This is also slightly different from the allow_port_security_disabled
option which is a temporary flag until liberty end of life at
which point nova can handle booting a server and attaching a
port with port security disabled.

To make this run in the integrated gate, devstack will need to
set CONF.network_feature_enabled.port_security=True in all
branches except stable/liberty.

Change-Id: I1efd5c838aa0d73cc6e8864e3041eea25850198d
Closes-Bug: #1624082