Deprecate nova api extensions config option
Now that the nova extensions mechanism is deprecated and api extensions
aren't really a thing anymore we should deprecate the corresponding
mechanism in tempest. Moving forward there really isn't anything to set
here because microversions are the api selection mechanism and the
"extension list" is fixed. We can remove this option when no supported
nova release contains extensions anymore.

Change-Id: I1286294b1b61aebf53437b37f6c141bc4c587b1f