Fill name param in create_test_server
When calling create_test_server without name specified,
the resouce name will be something like 'tempest.common.compute-
instance-244552854', and it is hard to distinguish the testcase
from such kind of name. This is to fill name in create_test_server
which uses testcase name as prefix to help debugging.

Change-Id: I4fbc1e984e833824b820421a03d99e3a475fad3a