Use other item instead of security_group in network-quotas test
In network/admin/test_quotas, it uses {'security_group': 0} for
new quotas, but if neutron's security_group extension is not
enabled, the testcase will fail unexpectedly.

As we are just testing neutron's quotas, we can choose quota items
that are common to all environments such as 'network' and 'port',
thus the testcase will work happily even if neutron's security-group
extension is not enabled.

Change-Id: I9147ea1c151672b25a25ceee6a541ddf7f04a8cd