Remove image v1 test_delete_image_with_invalid_image_id
This is a negative test that essentially ends up testing your http
server behavior because the url is not valid at all and http servers
catch that. This means that depending on your server you get different
results and tempest will fail.

Run into this when using apache to proxy tls for glance.

Suggestion is to remove the test for this reason.

Change-Id: I93c8e2c755dbbf66bada2031daa5e29b1488d2ea