Wait for floating_ip to detach from server after deletion
test_minimum_basic_scenario expects floating_ip detaching
from server immediately after deletion, and this may not
be true in production clouds(we watched in our system that
it took about 5 seconds for the floating_ip to disappear
from server's address). So this is to add wait after
floating_ip deletion.

Change-Id: Iedc6197829f2cb388878fdb8e257489482a41d93