Add list the primary tenant
s1 and s2 belong to primary tenant. We have tests that can not list
the primary tenant due to odd specified behavior and tests that
can not list other tenants. But readers will be confused that whether
s1 and s2 indeed belong to primary tenant? We do not test that we
can list primary tenant with all_tenants specified.  So I add this.

Change-Id: I8a6ec2b4994258fb530c9fba1ddbd8a48290e607