Use a common method for is a router interface
Tempest currently tries to figure out is port a router
interface or not by the device_owner of the port.

The logic in some places extended to be able to deal with dvr setups,
but in another places it was missing.

This change also adds support for L3-Ha support where
the interface owner is network:ha_router_replicated_interface.

By having common method it allows as to improve
the logic just by change it only in one place.

Change-Id: Ia83c8ad1bac9e44041b0661e26b7791a83087420