Use lib version of data_utils in dynamic creds
As part of the pending library migration for the dynamic credentials
module we need to remove all dependencies on non-lib code. As part of
this effort, this patch changes all usages of tempest's lib data_utils
wrapper to just use the lib version directly. The tempest wrapper only
sets a common prefix for rand_name, so this switches all uses of
rand_name to do the same.

Change-Id: I231c4522501249b9d6854f6378e75a0f130aea1c