Move InvalidConfiguration exception to tempest.lib
This commit migrates the InvalidConfiguartion exception from tempest's
exceptions into the list of tempest lib exceptions. While tempest lib
itself does not have any configuration, it is very possible that
external test suites would also like a way to raise an exception when
part of the configuration is invalid. This is also needed as part of
the migration of the dynamic credentials provider to tempest lib.

Change-Id: I9ec697c61cfb1f5750d82ae790c607029bfed3ca