Move cred_client to tempest.lib
This commit migrates the cred_client module from tempest.common to
tempest lib. This module provides an abstraction around the differences
between v2 credential operations and v3 credential operations. Since for
a lot of tests you just need to interact with credentials and which
version of keystone is deployed or being used doesn't matter. As part of
this migration unit tests are added to handle testing the abstraction
layer.

Change-Id: I196033483e06c96c16eaefeaac68b3253c3347f5