Skip test_volume_backed_live_migration unconditionally
The volume-backed live migration with iscsi has been failing
on ubuntu 14.04 and 16.04 in the gate for as long as it's
been tested. The current theory is there is a kernel issue
with qemu and iscsi and no one is really working on a fix, e.g.
testing qemu-native iscsi to see if that makes a difference.

The dedicated live-migration job has already been skipping this
since 2f68cd3e9f47ca5bb864c26c1b0b123918197ef4 and that stabilized
the job so now it's voting on nova changes.

The only reason the gate-tempest-dsvm-multinode* jobs are non-voting
is because of this test, so let's skip it for now. We might be
able to enable it with a different backing volume store like
ceph eventually but that's another work in progress.

Change-Id: Ibf067d15968f4f1fcf2deb8b6359a7e7ea7eb46e
Related-Bug: #1524898