Pass username on v3 token issue
The v3 test_create_token test was failing when no username was
specified in the accounts.yaml file. It was actually making an
invalid request rather than creating the request using the
username.

The underlying get_token method actually supports making a
request with username or password so looks like it was just a
mistake not passing the username.

Also, added support for unspecified fields in accounts.yaml.
Rather than fail because the server returns a user ID when
none is specified, just validate that a user ID is present
if no user ID is specified.

Change-Id: Ie91f615c16d4e898f4fe2bbe942303ba51418cbb