Fixing 'test_verify_created_server_ephemeral_disk' test
The 'test_verify_created_server_ephemeral_disk' test has been recently
broken by the patch [1]. The traceback can be found in logs of the
'gate-tempest-dsvm-neutron-full-ssh' job [2]. The test fails with error [3].

When we create a flavor with no ephemeral disk, the 'ephemeral' argument
is unexpected. So we don't need to pass the 'ephemeral' argument to the
'create_flavor' method when we create a flavor with no ephemeral disk.

[1] https://review.openstack.org/#/c/390198/
[2] http://logs.openstack.org/98/390198/2/check/gate-tempest-dsvm-neutron-full-ssh/fef3bfd/logs/testr_results.html.gz
[3] http://paste.openstack.org/show/587678/

Change-Id: I0cdbba827640d09d7b1fb8e1f3c0b56013a5d3a6