Add wait_for_server_termination in test_server_basic_ops
In test_server_basic_ops, the last step is "delete_server" and
without any further check. The pure "delete_server" is of no
use because there already exists addCleanup in self.create_server.

On the other hand, if we want to check whether the server can
be deleted properly, we should add wait_for_server_termination
after delete_server, thus it can raise an exception if the server
ends with unexpected ERROR state.

Change-Id: I2efa0f94faaa6adce5d9b559143456d060911125