Add sleep(2) when try to delete a container
Even though in devstack this test pass successfully, the same
don't happen in HA deployment. It takes a few seconds to
synchronize the deletion of the objects between the controllers
in order to delete the container. Without this time this test
fails consistently in tripleo as you can see in the logs at
http://bit.ly/2gMDl8d.
Adding the sleep makes the HA proxy have enough time to sync
the deletion of the objects, and the container is deleted
successfully.

Change-Id: I76f6493d1429166bef9804f2fbf2acdc02425489