Add test.related_bug() to know launchpad bug reports
We(QA-team) have added some tests into Tempest for reproducing bugs on the
gate and applied the corresponding patches into multiple branches. This kind
of tests are useful for production clouds also because they can know latent
bugs of their own clouds from the test failures.
However, it is difficult to find the corresponding patches when facing test
failures now because testers need to search the git history of the failed test
and find the bug number from "Related-Bug" tag from the commit message.
This patch adds a new decorator to know launchpad bug reports more easily.

In addition, this patch puts the decorator to test_delete_attached_volume
as a sample becauuse the test can detect the server internal error which
makes operations harder on productions.

Change-Id: Idc9ac0ebc61e230e4e014fb6476d1c74116511f7