Correct boundary value test in ImagesOneServerNegativeTestJSON
The max character length of name is 255 in create_image, and
we should use the length of 256 for the negative boundary value
test.
Besides, testcase name test_create_image_specify_name_over_256_chars
is not correct, in which 256 should be 255.

see http://developer.openstack.org/api-ref/image/v2/index.html?expanded=create-an-image-detail#create-an-image

Change-Id: I924ae3a386f9609a2ce3f6fafe368fa7290435d1