Wait for server to be active before deleting
In ListServersNegativeTestJSON's setup, the comment says,
"# We ignore errors on termination because the server may
 # be put into ERROR status on a quick spawn, then delete,
 # as the compute node expects the instance local status
 # to be spawning, not deleted.",
and it is true that deleting a server immediately when it
is spawning will sometimes lead to ERROR status, especially
in complex production clouds.

Instead of just ignoring the error, maybe we can use a more
positive way, i.e., to wait for server to be active before
deleting it, thus the server should be deleted successfully
most of the time(if server is in ERROR status, then something
must be wrong).

Through ERROR server could be deleted if we delete it again,
but current tempest code won't do twice deletion(usually it
is not necessary).

Change-Id: Iab6b4fef58c32d110f9454e49b4d7be8c9053fb6